[GH-248] Unity:raise error if service unavailable (#249)
When the Unity service is unavailable, getting any resource would return
UnityResourceNotFoundError. This behavior is wrong. For example,
OpenStack checks the host existing or not before creating a new one.

Let's say the Host-1 already exists, and if the GET request is sent when
the service unavailable, the UnityResourceNotFoundError will cause
another host created with the same name.