[#248] Unity: retry not work (#251)
Fix #249 to issue #248 is too open. It changed
`if resp.status_code == 401:` to
`if resp.status_code >= 400 and resp.status_code != 404:`, which means
status code like 422 (hlu number conflict) raise `HttpClientError` and
the `HttpClientError` cannot be caught afterward and exceptions to
Cinder. Thus the retry mechanism doesn't take effect.

Another example is that deleting a volume twice. The second would return
409, and `HttpClientError` is raised directly to Cinder. The expected
behavior should be that `ResourceNotFound` is caught and ignored by
Unity driver.